Reworked state classes to make more sense and have a clean API

+++ src/device.cpp
@@ -1,4 +1,5 @@
 #include "device.h"
+#include "device_record.h"
 #include "user_record.h"
 
 void device::delete_user_record(uint64_t user_index) {
@@ -8,24 +9,21 @@ void device::delete_user_record(uint64_t user_index) {
     }
 }
 
-void device::delete_user_record(const user_record& ur) {
-    auto user_index = std::find_if(correspondents.begin(), correspondents.end(), [&](const auto& p){return (p.second == ur);})->first;
-    return delete_user_record(user_index);
-}
-
-void device::delete_device_record(user_record& ur, const device_record& dr) {
-    if (!ur.remove_device(dr)) {
-        //Need to delete user record as well
-        delete_user_record(ur);
+void device::delete_device_record(uint64_t user_index, uint64_t device_index) {
+    auto user_rec = correspondents.at(user_index);
+    if (user_rec.delete_device_record(device_index)) {
+        delete_user_record(user_index);
     }
 }
 
-void device::delete_session(user_record& ur, device_record& dr, const session& s) {
-    if (!dr.delete_session(s)) {
-        delete_device_record(ur, dr);
+void device::delete_session(uint64_t user_index, uint64_t device_index, const session& s) {
+    auto device_rec = correspondents.at(user_index).user_devices.at(device_index);
+    if (device_rec.delete_session(s)) {
+        delete_device_record(user_index, device_index);
     }
 }
 
-void device::insert_session(device_record& dr, const session& s) {
-    dr.insert_session(s);
+void device::insert_session(uint64_t user_index, uint64_t device_index, const session& s) {
+    auto device_rec = correspondents.at(user_index).user_devices.at(device_index);
+    device_rec.insert_session(s);
 }
